Remove base url from index
It was added to try to make the CI work.
But it breaks local development (has the js bundle downloaded
is the one from the prod)
Also now that I've changed the links to relative ones (`./[...]`)
I don't think it's necessary.

Worth case scenario: I'll need to set it up dynamically
depending on the environment